+++ packages/bridge-controller/src/utils/fetch.test.ts
@@ -1,4 +1,4 @@
-import { ZeroAddress } from 'ethers';
+import { AddressZero } from '@ethersproject/constants';
 
 import {
   fetchBridgeFeatureFlags,
@@ -7,7 +7,7 @@ import {
 } from './fetch';
 import mockBridgeQuotesErc20Erc20 from '../../tests/mock-quotes-erc20-erc20.json';
 import mockBridgeQuotesNativeErc20 from '../../tests/mock-quotes-native-erc20.json';
-import { BridgeClientId } from '../constants/bridge';
+import { BridgeClientId, BRIDGE_PROD_API_BASE_URL } from '../constants/bridge';
 import { CHAIN_IDS } from '../constants/chains';
 
 const mockFetchFn = jest.fn();
@@ -56,6 +56,7 @@ describe('fetch', () => {
       const result = await fetchBridgeFeatureFlags(
         BridgeClientId.EXTENSION,
         mockFetchFn,
+        BRIDGE_PROD_API_BASE_URL,
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -129,6 +130,7 @@ describe('fetch', () => {
       const result = await fetchBridgeFeatureFlags(
         BridgeClientId.EXTENSION,
         mockFetchFn,
+        BRIDGE_PROD_API_BASE_URL,
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -156,7 +158,11 @@ describe('fetch', () => {
       mockFetchFn.mockRejectedValue(mockError);
 
       await expect(
-        fetchBridgeFeatureFlags(BridgeClientId.EXTENSION, mockFetchFn),
+        fetchBridgeFeatureFlags(
+          BridgeClientId.EXTENSION,
+          mockFetchFn,
+          BRIDGE_PROD_API_BASE_URL,
+        ),
       ).rejects.toThrow(mockError);
     });
   });
@@ -196,6 +202,7 @@ describe('fetch', () => {
         '0xa',
         BridgeClientId.EXTENSION,
         mockFetchFn,
+        BRIDGE_PROD_API_BASE_URL,
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -233,7 +240,12 @@ describe('fetch', () => {
       mockFetchFn.mockRejectedValue(mockError);
 
       await expect(
-        fetchBridgeTokens('0xa', BridgeClientId.EXTENSION, mockFetchFn),
+        fetchBridgeTokens(
+          '0xa',
+          BridgeClientId.EXTENSION,
+          mockFetchFn,
+          BRIDGE_PROD_API_BASE_URL,
+        ),
       ).rejects.toThrow(mockError);
     });
   });
@@ -248,14 +260,15 @@ describe('fetch', () => {
           walletAddress: '0x123',
           srcChainId: 1,
           destChainId: 10,
-          srcTokenAddress: ZeroAddress,
-          destTokenAddress: ZeroAddress,
+          srcTokenAddress: AddressZero,
+          destTokenAddress: AddressZero,
           srcTokenAmount: '20000',
           slippage: 0.5,
         },
         signal,
         BridgeClientId.EXTENSION,
         mockFetchFn,
+        BRIDGE_PROD_API_BASE_URL,
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -282,14 +295,15 @@ describe('fetch', () => {
           walletAddress: '0x123',
           srcChainId: 1,
           destChainId: 10,
-          srcTokenAddress: ZeroAddress,
-          destTokenAddress: ZeroAddress,
+          srcTokenAddress: AddressZero,
+          destTokenAddress: AddressZero,
           srcTokenAmount: '20000',
           slippage: 0.5,
         },
         signal,
         BridgeClientId.EXTENSION,
         mockFetchFn,
+        BRIDGE_PROD_API_BASE_URL,
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -335,14 +349,15 @@ describe('fetch', () => {
           walletAddress: '0x123',
           srcChainId: 1,
           destChainId: 10,
-          srcTokenAddress: ZeroAddress,
-          destTokenAddress: ZeroAddress,
+          srcTokenAddress: AddressZero,
+          destTokenAddress: AddressZero,
           srcTokenAmount: '20000',
           slippage: 0.5,
         },
         signal,
         BridgeClientId.EXTENSION,
         mockFetchFn,
+        BRIDGE_PROD_API_BASE_URL,
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(