+++ packages/bridge-controller/src/utils/fetch.ts
@@ -4,32 +4,23 @@ import { hexToNumber, numberToHex } from '@metamask/utils';
 import {
   isSwapsDefaultTokenAddress,
   isSwapsDefaultTokenSymbol,
-  getBridgeApiBaseUrl,
 } from './bridge';
 import {
-  FEATURE_FLAG_VALIDATORS,
-  QUOTE_VALIDATORS,
-  TX_DATA_VALIDATORS,
-  TOKEN_VALIDATORS,
-  validateResponse,
-  QUOTE_RESPONSE_VALIDATORS,
-  FEE_DATA_VALIDATORS,
+  validateFeatureFlagsResponse,
+  validateQuoteResponse,
+  validateSwapsTokenObject,
 } from './validators';
 import { DEFAULT_FEATURE_FLAG_CONFIG } from '../constants/bridge';
 import type { SwapsTokenObject } from '../constants/tokens';
 import { SWAPS_CHAINID_DEFAULT_TOKEN_MAP } from '../constants/tokens';
 import type {
-  FeatureFlagResponse,
-  FeeData,
-  Quote,
   QuoteRequest,
   QuoteResponse,
-  TxData,
   BridgeFeatureFlags,
   FetchFunction,
   ChainConfiguration,
 } from '../types';
-import { BridgeFlag, FeeType, BridgeFeatureFlagsKey } from '../types';
+import { BridgeFlag, BridgeFeatureFlagsKey } from '../types';
 
 // TODO put this back in once we have a fetchWithCache equivalent
 // const CACHE_REFRESH_TEN_MINUTES = 10 * Duration.Minute;
@@ -43,24 +34,20 @@ export const getClientIdHeader = (clientId: string) => ({
  *
  * @param clientId - The client ID for metrics
  * @param fetchFn - The fetch function to use
+ * @param bridgeApiBaseUrl - The base URL for the bridge API
  * @returns The bridge feature flags
  */
 export async function fetchBridgeFeatureFlags(
   clientId: string,
   fetchFn: FetchFunction,
+  bridgeApiBaseUrl: string,
 ): Promise<BridgeFeatureFlags> {
-  const url = `${getBridgeApiBaseUrl()}/getAllFeatureFlags`;
-  const rawFeatureFlags = await fetchFn(url, {
+  const url = `${bridgeApiBaseUrl}/getAllFeatureFlags`;
+  const rawFeatureFlags: unknown = await fetchFn(url, {
     headers: getClientIdHeader(clientId),
   });
 
-  if (
-    validateResponse<FeatureFlagResponse>(
-      FEATURE_FLAG_VALIDATORS,
-      rawFeatureFlags,
-      url,
-    )
-  ) {
+  if (validateFeatureFlagsResponse(rawFeatureFlags)) {
     const getChainsObj = (chains: Record<number, ChainConfiguration>) =>
       Object.entries(chains).reduce(
         (acc, [chainId, value]) => ({
@@ -96,17 +83,17 @@ export async function fetchBridgeFeatureFlags(
  * @param chainId - The chain ID to fetch tokens for
  * @param clientId - The client ID for metrics
  * @param fetchFn - The fetch function to use
+ * @param bridgeApiBaseUrl - The base URL for the bridge API
  * @returns A list of enabled (unblocked) tokens
  */
 export async function fetchBridgeTokens(
   chainId: Hex,
   clientId: string,
   fetchFn: FetchFunction,
+  bridgeApiBaseUrl: string,
 ): Promise<Record<string, SwapsTokenObject>> {
   // TODO make token api v2 call
-  const url = `${getBridgeApiBaseUrl()}/getTokens?chainId=${hexToNumber(
-    chainId,
-  )}`;
+  const url = `${bridgeApiBaseUrl}/getTokens?chainId=${hexToNumber(chainId)}`;
 
   // TODO we will need to cache these. In Extension fetchWithCache is used. This is due to the following:
   // If we allow selecting dest networks which the user has not imported,
@@ -127,7 +114,7 @@ export async function fetchBridgeTokens(
 
   tokens.forEach((token: unknown) => {
     if (
-      validateResponse<SwapsTokenObject>(TOKEN_VALIDATORS, token, url, false) &&
+      validateSwapsTokenObject(token) &&
       !(
         isSwapsDefaultTokenSymbol(token.symbol, chainId) ||
         isSwapsDefaultTokenAddress(token.address, chainId)
@@ -146,6 +133,7 @@ export async function fetchBridgeTokens(
  * @param signal - The abort signal
  * @param clientId - The client ID for metrics
  * @param fetchFn - The fetch function to use
+ * @param bridgeApiBaseUrl - The base URL for the bridge API
  * @returns A list of bridge tx quotes
  */
 export async function fetchBridgeQuotes(
@@ -153,6 +141,7 @@ export async function fetchBridgeQuotes(
   signal: AbortSignal,
   clientId: string,
   fetchFn: FetchFunction,
+  bridgeApiBaseUrl: string,
 ): Promise<QuoteResponse[]> {
   const queryParams = new URLSearchParams({
     walletAddress: request.walletAddress,
@@ -165,41 +154,14 @@ export async function fetchBridgeQuotes(
     insufficientBal: request.insufficientBal ? 'true' : 'false',
     resetApproval: request.resetApproval ? 'true' : 'false',
   });
-  const url = `${getBridgeApiBaseUrl()}/getQuote?${queryParams}`;
-  const quotes = await fetchFn(url, {
+  const url = `${bridgeApiBaseUrl}/getQuote?${queryParams}`;
+  const quotes: unknown[] = await fetchFn(url, {
     headers: getClientIdHeader(clientId),
     signal,
   });
 
-  const filteredQuotes = quotes.filter((quoteResponse: QuoteResponse) => {
-    const { quote, approval, trade } = quoteResponse;
-    return (
-      validateResponse<QuoteResponse>(
-        QUOTE_RESPONSE_VALIDATORS,
-        quoteResponse,
-        url,
-      ) &&
-      validateResponse<Quote>(QUOTE_VALIDATORS, quote, url) &&
-      validateResponse<SwapsTokenObject>(
-        TOKEN_VALIDATORS,
-        quote.srcAsset,
-        url,
-      ) &&
-      validateResponse<SwapsTokenObject>(
-        TOKEN_VALIDATORS,
-        quote.destAsset,
-        url,
-      ) &&
-      validateResponse<TxData>(TX_DATA_VALIDATORS, trade, url) &&
-      validateResponse<FeeData>(
-        FEE_DATA_VALIDATORS,
-        quote.feeData[FeeType.METABRIDGE],
-        url,
-      ) &&
-      (approval
-        ? validateResponse<TxData>(TX_DATA_VALIDATORS, approval, url)
-        : true)
-    );
+  const filteredQuotes = quotes.filter((quoteResponse: unknown) => {
+    return validateQuoteResponse(quoteResponse);
   });
-  return filteredQuotes;
+  return filteredQuotes as QuoteResponse[];
 }