+++ packages/transaction-controller/src/utils/eip7702.test.ts
@@ -4,7 +4,6 @@ import type { RemoteFeatureFlagControllerGetStateAction } from '@metamask/remote
 import type { Hex } from '@metamask/utils';
 import { remove0x } from '@metamask/utils';
 
-import type { KeyringControllerSignAuthorization } from './eip7702';
 import {
   DELEGATION_PREFIX,
   doesChainSupportEIP7702,
@@ -17,6 +16,7 @@ import {
   getEIP7702SupportedChains,
 } from './feature-flags';
 import { Messenger } from '../../../base-controller/src';
+import type { KeyringControllerSignEip7702AuthorizationAction } from '../../../keyring-controller/src';
 import type { TransactionControllerMessenger } from '../TransactionController';
 import type { AuthorizationList } from '../types';
 import { TransactionStatus, type TransactionMeta } from '../types';
@@ -33,6 +33,7 @@ const CHAIN_ID_2_MOCK = '0x456';
 const ADDRESS_MOCK = '0x1234567890123456789012345678901234567890';
 const ADDRESS_2_MOCK = '0x0987654321098765432109876543210987654321';
 const ADDRESS_3_MOCK = '0xabcdefabcdefabcdefabcdefabcdefabcdefabcd';
+const PUBLIC_KEY_MOCK = '0x112233';
 const ETH_QUERY_MOCK = {} as EthQuery;
 
 const DATA_MOCK =
@@ -72,7 +73,7 @@ const AUTHORIZATION_LIST_MOCK: AuthorizationList = [
 
 describe('EIP-7702 Utils', () => {
   let baseMessenger: Messenger<
-    | KeyringControllerSignAuthorization
+    | KeyringControllerSignEip7702AuthorizationAction
     | RemoteFeatureFlagControllerGetStateAction,
     never
   >;
@@ -87,7 +88,7 @@ describe('EIP-7702 Utils', () => {
   );
 
   let signAuthorizationMock: jest.MockedFn<
-    KeyringControllerSignAuthorization['handler']
+    KeyringControllerSignEip7702AuthorizationAction['handler']
   >;
 
   beforeEach(() => {
@@ -100,13 +101,13 @@ describe('EIP-7702 Utils', () => {
       .mockResolvedValue(AUTHORIZATION_SIGNATURE_MOCK);
 
     baseMessenger.registerActionHandler(
-      'KeyringController:signEip7702AuthorizationMessage',
+      'KeyringController:signEip7702Authorization',
       signAuthorizationMock,
     );
 
     controllerMessenger = baseMessenger.getRestricted({
       name: 'TransactionController',
-      allowedActions: ['KeyringController:signEip7702AuthorizationMessage'],
+      allowedActions: ['KeyringController:signEip7702Authorization'],
       allowedEvents: [],
     });
   });
@@ -271,6 +272,7 @@ describe('EIP-7702 Utils', () => {
         await isAccountUpgradedToEIP7702(
           ADDRESS_MOCK,
           CHAIN_ID_MOCK,
+          PUBLIC_KEY_MOCK,
           controllerMessenger,
           ETH_QUERY_MOCK,
         ),
@@ -293,6 +295,7 @@ describe('EIP-7702 Utils', () => {
         await isAccountUpgradedToEIP7702(
           ADDRESS_MOCK,
           CHAIN_ID_MOCK.toUpperCase() as Hex,
+          PUBLIC_KEY_MOCK,
           controllerMessenger,
           ETH_QUERY_MOCK,
         ),
@@ -313,6 +316,7 @@ describe('EIP-7702 Utils', () => {
         await isAccountUpgradedToEIP7702(
           ADDRESS_MOCK,
           CHAIN_ID_MOCK,
+          PUBLIC_KEY_MOCK,
           controllerMessenger,
           ETH_QUERY_MOCK,
         ),
@@ -331,6 +335,7 @@ describe('EIP-7702 Utils', () => {
         await isAccountUpgradedToEIP7702(
           ADDRESS_MOCK,
           CHAIN_ID_MOCK,
+          PUBLIC_KEY_MOCK,
           controllerMessenger,
           ETH_QUERY_MOCK,
         ),
@@ -349,6 +354,7 @@ describe('EIP-7702 Utils', () => {
         await isAccountUpgradedToEIP7702(
           ADDRESS_MOCK,
           CHAIN_ID_MOCK,
+          PUBLIC_KEY_MOCK,
           controllerMessenger,
           ETH_QUERY_MOCK,
         ),
@@ -369,6 +375,7 @@ describe('EIP-7702 Utils', () => {
         await isAccountUpgradedToEIP7702(
           ADDRESS_MOCK,
           CHAIN_ID_MOCK,
+          PUBLIC_KEY_MOCK,
           controllerMessenger,
           ETH_QUERY_MOCK,
         ),